Update: Code Generator + Instrumentor
Code Generator now automatically makes sure new components are available in the editor drop down. It feels a little bit hacky personally but I'm not sure how else to do it. I'm implementing a static function reflect::InitApplicationReflection per app and this is defined via ReflectedData_gen.cpp which the FoundationCodeGenerator makes. So any code inside this function will be run in the application constructor. For now, I'm just calling AddComponent on any component I find during code generation so that the TypeDescriptor will automatically be called and therefore initReflection will be called, thus adding the Component to the ClassRegistry (which is what the AddComponent button loops over in editor)

Added Instrumentor again, gets ignored because it's under a folder named Debug. Changed this to be called Profiling so any more Profiling tools can be placed here too and won't be ignored going forward.