Update: Scripts
Cleaning up scripts folder so it's divided into BuildMachine (scripts that are only intended to be run on a BuildMachine external from the local machine). Local, scripts intended to be run on each users machine. Setup.bat is left where it is for ease of use and to be referenced in the README.md (when I actually update it with building instructions.)