Update: Serialization
Making sure we set the owners of class objects to whatever base object we are passing in.

This was causing a bug when the user tries to remove a component from the editor panel after loading in a level. This is because the component didn't have any owner and we need to know who the owner is in order to remove the component from them.